correctQ/backend/requirements.txt
fastapi>=0.115.12
python-dotenv>=1.1.0
sqlalchemy>=2.0.41
svix>=1.65.0
uvicorn>=0.34.2
google-genai
supabase
pymongo
python-dotenv
python-multipart